+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/factory/KakaoPayRequestFactory.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.application;
+package com.boeingmerryho.business.paymentservice.application.factory;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
