+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/service/PaySessionService.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.application;
+package com.boeingmerryho.business.paymentservice.application.service;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
