+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/service/PaymentFeignService.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.application;
+package com.boeingmerryho.business.paymentservice.application.service;
 
 import org.springframework.transaction.annotation.Transactional;
