+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/strategy/PaymentStrategy.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.application;
+package com.boeingmerryho.business.paymentservice.application.strategy;
 
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,8 +12,8 @@ import com.boeingmerryho.business.paymentservice.application.dto.response.Paymen
 import com.boeingmerryho.business.paymentservice.domain.entity.Payment;
 import com.boeingmerryho.business.paymentservice.domain.entity.PaymentDetail;
 import com.boeingmerryho.business.paymentservice.domain.type.PaymentMethod;
-import com.boeingmerryho.business.paymentservice.infrastructure.exception.ErrorCode;
 import com.boeingmerryho.business.paymentservice.infrastructure.exception.PaymentException;
+import com.boeingmerryho.business.paymentservice.presentation.code.PaymentErrorCode;
 
 public interface PaymentStrategy {
 	PaymentMethod getSupportedMethod();
@@ -29,14 +29,14 @@ public interface PaymentStrategy {
 		PaymentSession paymentSession,
 		Payment payment,
 		PaymentApproveRequestServiceDto requestDto) {
-		throw new PaymentException(ErrorCode.PAYMENT_UNSUPPORTED);
+		throw new PaymentException(PaymentErrorCode.PAYMENT_UNSUPPORTED);
 	}
 
 	@Transactional
 	default PaymentApproveResponseServiceDto approve(
 		Payment payment,
 		PaymentApproveAdminRequestServiceDto requestDto) {
-		throw new PaymentException(ErrorCode.PAYMENT_UNSUPPORTED);
+		throw new PaymentException(PaymentErrorCode.PAYMENT_UNSUPPORTED);
 	}
 
 	@Transactional
