+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/domain/helper/PaymentReader.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.domain;
+package com.boeingmerryho.business.paymentservice.domain.helper;
 
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Component;
@@ -8,8 +8,8 @@ import com.boeingmerryho.business.paymentservice.domain.entity.Payment;
 import com.boeingmerryho.business.paymentservice.domain.entity.PaymentDetail;
 import com.boeingmerryho.business.paymentservice.domain.repository.PaymentDetailRepository;
 import com.boeingmerryho.business.paymentservice.domain.repository.PaymentRepository;
-import com.boeingmerryho.business.paymentservice.infrastructure.exception.ErrorCode;
 import com.boeingmerryho.business.paymentservice.infrastructure.exception.PaymentException;
+import com.boeingmerryho.business.paymentservice.presentation.code.PaymentErrorCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,21 +24,21 @@ public class PaymentReader {
 		Long paymentId
 	) {
 		return paymentRepository.findById(paymentId)
-			.orElseThrow(() -> new PaymentException(ErrorCode.PAYMENT_NOT_FOUND));
+			.orElseThrow(() -> new PaymentException(PaymentErrorCode.PAYMENT_NOT_FOUND));
 	}
 
 	public PaymentDetail getDetail(
 		Long detailId
 	) {
 		return paymentDetailRepository.findById(detailId)
-			.orElseThrow(() -> new PaymentException(ErrorCode.PAYMENT_DETAIL_NOT_FOUND));
+			.orElseThrow(() -> new PaymentException(PaymentErrorCode.PAYMENT_DETAIL_NOT_FOUND));
 	}
 
 	public PaymentDetail getDetailByPaymentId(
 		Long paymentId
 	) {
 		return paymentDetailRepository.findPaymentDetailByPaymentId(paymentId)
-			.orElseThrow(() -> new PaymentException(ErrorCode.PAYMENT_DETAIL_NOT_FOUND));
+			.orElseThrow(() -> new PaymentException(PaymentErrorCode.PAYMENT_DETAIL_NOT_FOUND));
 	}
 
 	public Page<PaymentDetail> getPaymentDetails(
