+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/infrastructure/helper/KafkaProducerHelper.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.infrastructure;
+package com.boeingmerryho.business.paymentservice.infrastructure.helper;
 
 import java.util.List;
