+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/infrastructure/service/KakaoPayStrategy.java
@@ -9,9 +9,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
-import com.boeingmerryho.business.paymentservice.application.KakaoPayRequestFactory;
-import com.boeingmerryho.business.paymentservice.application.PaymentSessionFactory;
-import com.boeingmerryho.business.paymentservice.application.PaymentStrategy;
 import com.boeingmerryho.business.paymentservice.application.dto.PaymentApplicationMapper;
 import com.boeingmerryho.business.paymentservice.application.dto.kakao.KakaoPayApproveRequest;
 import com.boeingmerryho.business.paymentservice.application.dto.kakao.KakaoPayApproveResponse;
@@ -25,18 +22,21 @@ import com.boeingmerryho.business.paymentservice.application.dto.request.Payment
 import com.boeingmerryho.business.paymentservice.application.dto.response.PaymentApproveResponseServiceDto;
 import com.boeingmerryho.business.paymentservice.application.dto.response.PaymentReadyResponseServiceDto;
 import com.boeingmerryho.business.paymentservice.application.dto.response.PaymentRefundResponseServiceDto;
-import com.boeingmerryho.business.paymentservice.domain.PaymentFactory;
+import com.boeingmerryho.business.paymentservice.application.factory.KakaoPayRequestFactory;
+import com.boeingmerryho.business.paymentservice.application.factory.PaymentSessionFactory;
+import com.boeingmerryho.business.paymentservice.application.strategy.PaymentStrategy;
 import com.boeingmerryho.business.paymentservice.domain.entity.Payment;
 import com.boeingmerryho.business.paymentservice.domain.entity.PaymentDetail;
 import com.boeingmerryho.business.paymentservice.domain.entity.PaymentMembership;
 import com.boeingmerryho.business.paymentservice.domain.entity.PaymentTicket;
+import com.boeingmerryho.business.paymentservice.domain.factory.PaymentFactory;
 import com.boeingmerryho.business.paymentservice.domain.repository.PaymentDetailRepository;
 import com.boeingmerryho.business.paymentservice.domain.repository.PaymentRepository;
 import com.boeingmerryho.business.paymentservice.domain.type.PaymentMethod;
 import com.boeingmerryho.business.paymentservice.domain.type.PaymentType;
-import com.boeingmerryho.business.paymentservice.infrastructure.KafkaProducerHelper;
-import com.boeingmerryho.business.paymentservice.infrastructure.KakaoApiClient;
-import com.boeingmerryho.business.paymentservice.infrastructure.PaySessionHelper;
+import com.boeingmerryho.business.paymentservice.infrastructure.helper.KafkaProducerHelper;
+import com.boeingmerryho.business.paymentservice.infrastructure.helper.KakaoApiClient;
+import com.boeingmerryho.business.paymentservice.infrastructure.helper.PaySessionHelper;
 import com.boeingmerryho.business.paymentservice.presentation.dto.request.Ticket;
 
 import lombok.RequiredArgsConstructor;
