+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/presentation/code/PaymentSuccessCode.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.presentation;
+package com.boeingmerryho.business.paymentservice.presentation.code;
 
 import org.springframework.http.HttpStatus;
